requirements.txt
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.5.3
PyMySQL==0.7.11
PySocks==1.6.7
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
urllib3==1.22
Werkzeug==0.12.2
